+++ gcovr/tests/cmake_gtest/source/code.cpp
@@ -1,0 +1,18 @@
+#include "code.hpp"
+
+int foo1(int param) {
+  if (param) {
+     return 1;
+  } else {
+     return 0;
+  }
+}
+
+
+int foo2(int param) {
+  if (param) {
+     return 1;
+  } else {
+     return 0;
+  }
+}