+++ gcovr/tests/cmake_oos_ninja/Toolchain.cmake
@@ -1,7 +1,0 @@
-set(CMAKE_SYSTEM_NAME GNU)
-
-set(CMAKE_C_COMPILER $ENV{CC})
-set(CMAKE_CXX_COMPILER $ENV{CXX})
-
-add_compile_options( --coverage )
-add_link_options( --coverage )