+++ gcovr/tests/source_from_pipe/test.cpp
@@ -1,0 +1,6 @@
+#include "code.hpp"
+
+int main(int argc, char** argv)
+{
+	return foo1(foo2(1) - 1);
+}