+++ noxfile.py
@@ -20,10 +20,13 @@
 import glob
 import os
 import platform
+import re
 import shutil
 import shlex
+import subprocess
 import sys
 import nox
+
 
 GCC_VERSIONS = [
     "gcc-5",
@@ -35,13 +38,68 @@ GCC_VERSIONS = [
     "clang-10",
     "clang-13",
 ]
-GCC_VERSION2USE = os.path.split(os.environ.get("CC", "gcc-5"))[1]
+
+GCC_VERSIONS_NEWEST_FIRST = [
+    "-".join(cc)
+    for cc in sorted(
+        [(*cc.split("-"),) for cc in GCC_VERSIONS],
+        key=lambda cc: (cc[0], int(cc[1])),
+        reverse=True,
+    )
+]
+
 DEFAULT_TEST_DIRECTORIES = ["doc", "gcovr"]
-DEFAULT_LINT_ARGUMENTS = ["setup.py", "noxfile.py", "admin"] + DEFAULT_TEST_DIRECTORIES
+DEFAULT_LINT_ARGUMENTS = [
+    "setup.py",
+    "noxfile.py",
+    "scripts",
+    "admin",
+] + DEFAULT_TEST_DIRECTORIES
 
 BLACK_PINNED_VERSION = "black==22.3.0"
 
 nox.options.sessions = ["qa"]
+
+
+def get_gcc_version_to_use():
+    # If the user explicitly set CC variable, use that directly without checks.
+    cc = os.environ.get("CC")
+    if cc:
+        return os.path.split(cc)[1]
+
+    # Find the first insalled compiler version we suport
+    for cc in GCC_VERSIONS_NEWEST_FIRST:
+        if shutil.which(cc):
+            return cc
+
+    for cc in ["gcc", "clang"]:
+        output = subprocess.check_output([cc, "--version"]).decode()
+        # Ignore error code since we want to find a valid executable
+
+        # look for a line "gcc WHATEVER VERSION.WHATEVER" in output like:
+        #    gcc (Ubuntu 9.4.0-1ubuntu1~20.04.1) 9.4.0
+        #    Copyright (C) 2019 Free Software Foundation, Inc.
+        #    This is free software; see the source for copying conditions.  There is NO
+        #    warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+        search_gcc_version = re.search(r"^gcc\b.* ([0-9]+)\.\S+$", output, re.M)
+
+        # look for a line "WHATEVER clang version VERSION.WHATEVER" in output like:
+        #    Apple clang version 13.1.6 (clang-1316.0.21.2.5)
+        #    Target: arm64-apple-darwin21.5.0
+        #    Thread model: posix
+        #    InstalledDir: /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin
+        search_clang_version = re.search(r"\bclang version ([0-9]+)\.", output, re.M)
+
+        if search_gcc_version:
+            major_version = search_gcc_version.group(1)
+            return f"gcc-{major_version}"
+        elif search_clang_version:
+            major_version = search_clang_version.group(1)
+            return f"clang-{major_version}"
+
+    raise RuntimeError(
+        "Could not detect a valid compiler, you can defin one by setting the environment CC"
+    )
 
 
 def set_environment(session: nox.Session, cc: str, check: bool = True) -> None:
@@ -59,7 +117,7 @@ def set_environment(session: nox.Session, cc: str, check: bool = True) -> None:
 @nox.session(python=False)
 def qa(session: nox.Session) -> None:
     """Run the quality tests for the default GCC version."""
-    session_id = f"qa_compiler({GCC_VERSION2USE})"
+    session_id = f"qa_compiler({get_gcc_version_to_use()})"
     session.log(f"Notify session {session_id}")
     session.notify(session_id)
 
@@ -129,7 +187,7 @@ def doc(session: nox.Session) -> None:
 @nox.session(python=False)
 def tests(session: nox.Session) -> None:
     """Run the tests with the default GCC version."""
-    session_id = f"tests_compiler({GCC_VERSION2USE})"
+    session_id = f"tests_compiler({get_gcc_version_to_use()})"
     session.log(f"Notify session {session_id}")
     session.notify(session_id)
 
@@ -146,7 +204,7 @@ def tests_compiler_all(session: nox.Session) -> None:
 @nox.session
 @nox.parametrize("version", [nox.param(v, id=v) for v in GCC_VERSIONS])
 def tests_compiler(session: nox.Session, version: str) -> None:
-    """Run the test with a specifiv GCC version."""
+    """Run the test with a specific GCC version."""
     session.install(
         "jinja2",
         "lxml",
@@ -201,17 +259,17 @@ def build_wheel(session: nox.Session) -> None:
     if os.path.isdir(dist_cache):
         shutil.rmtree(dist_cache)
     shutil.copytree("dist", dist_cache)
-    session.notify("check_wheel")
 
 
 @nox.session
 def check_wheel(session: nox.Session) -> None:
-    """Check the wheel, should not be used directly."""
+    """Check the wheel and do a smoke test, should not be used directly."""
     session.install("wheel", "twine")
     session.chdir(f"{session.cache_dir}/dist")
     session.run("twine", "check", "*", external=True)
     session.install(glob.glob("*.whl")[0])
     session.run("python", "-m", "gcovr", "--help", external=True)
+    session.run("gcovr", "--help", external=True)
 
 
 @nox.session
@@ -219,6 +277,46 @@ def upload_wheel(session: nox.Session) -> None:
     """Upload the wheel."""
     session.install("twine")
     session.run("twine", "upload", "dist/*", external=True)
+
+
+@nox.session
+def bundle_app(session: nox.Session) -> None:
+    """Bundle a standalone executable."""
+    session.install("pyinstaller")
+    session.install("-e", ".")
+    os.makedirs("build", exist_ok=True)
+    session.chdir("build")
+    if platform.system() == "Windows":
+        executable = "gcovr.exe"
+    else:
+        executable = "gcovr"
+    session.run(
+        "pyinstaller",
+        "--distpath",
+        ".",
+        "--workpath",
+        "./pyinstaller",
+        "--specpath",
+        "./pyinstaller",
+        "--onefile",
+        "--collect-all",
+        "gcovr",
+        "-n",
+        executable,
+        *session.posargs,
+        "../scripts/pyinstaller_entrypoint.py",
+    )
+    session.notify("check_bundled_app")
+
+
+@nox.session
+def check_bundled_app(session: nox.Session) -> None:
+    """Run a smoke test with the bundled app, should not be used directly."""
+    session.chdir("build")
+    session.run("bash", "-c", "./gcovr --help", external=True)
+    session.log("Run HTML all transformations to check if all the modules are packed")
+    for format in ["txt", "html", "cobertura", "sonarqube", "csv", "coveralls"]:
+        session.run("bash", "-c", f"./gcovr --{format} out.{format}", external=True)
 
 
 def docker_container_os(session: nox.Session) -> str:
@@ -237,7 +335,7 @@ def docker_container_id(session: nox.Session, version: str) -> str:
 @nox.session(python=False)
 def docker_qa_build(session: nox.Session) -> None:
     """Build the docker container for the default GCC version."""
-    session_id = f"docker_qa_build({GCC_VERSION2USE})"
+    session_id = f"docker_qa_build({GCC_VERSIONS[0]})"
     session.log(f"Notify session {session_id}")
     session.notify(session_id)
 
@@ -279,7 +377,7 @@ def docker_qa_build_compiler(session: nox.Session, version: str) -> None:
 @nox.session(python=False)
 def docker_qa_run(session: nox.Session) -> None:
     """Run the docker container for the default GCC version."""
-    session_id = f"docker_qa_run_compiler({GCC_VERSION2USE})"
+    session_id = f"docker_qa_run_compiler({GCC_VERSIONS[0]})"
     session.log(f"Notify session {session_id}")
     session.notify(session_id)
 
@@ -333,7 +431,7 @@ def docker_qa_run_compiler(session: nox.Session, version: str) -> None:
 @nox.session(python=False)
 def docker_qa(session: nox.Session) -> None:
     """Build and run the docker container for the default GCC version."""
-    session_id = f"docker_qa_compiler({GCC_VERSION2USE})"
+    session_id = f"docker_qa_compiler({GCC_VERSIONS[0]})"
     session.log(f"Notify session {session_id}")
     session.notify(session_id)